split precompiles into smaller, more readable files
